Wrap bottom navigation in Material widget
The bottom navigation bar is now wrapped in a Material widget to provide
elevation and better visual separation from the content above. This
improves the UI by adding depth and ensuring the navigation bar stands
out against the main content area. The elevation is set to 8 for a
subtle shadow effect.

Signed-off-by: Eric Curtin <eric.curtin@docker.com>